Remove the CAN_INSTALL file when occ maintenance:install is complete
When occ maintenance:install is run from the CLI, the CAN_INSTALL in the config directory is left in place when installed is set to true, whereas this file is removed when the web installer is used.  Removing this file in the CLI command maintains consistency between the two.  This allows automation tools an easier way to determine if this process has been completed.

Signed-off-by: Alex Harpin <development@landsofshadow.co.uk>